Session Master Basic impl done.

+++ src/main/java/com/vistaluxhms/repository/SessionRepository.java
@@ -5,6 +5,6 @@ import com.vistaluxhms.entity.SessionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface SessionRepository extends JpaRepository<SessionEntity, Long>, JpaSpecificationExecutor<SessionEntity> {
+public interface SessionRepository extends JpaRepository<SessionEntity, Integer>, JpaSpecificationExecutor<SessionEntity> {
 
 }
